Break out matcher.hpp
