+++ include/trompeloeil/matcher/compare.hpp
@@ -15,8 +15,8 @@
 #ifndef TROMPELOEIL_COMPARE_HPP
 #define TROMPELOEIL_COMPARE_HPP
 
-#ifndef TROMPELOEIL_MOCK_HPP_
-#include "../mock.hpp"
+#ifndef TROMPELOEIL_MATCHER_HPP
+#include "../matcher.hpp"
 #endif
 
 namespace trompeloeil
