+++ include/trompeloeil/matcher/deref.hpp
@@ -16,8 +16,8 @@
 #ifndef TROMPELOEIL_DEREF_HPP
 #define TROMPELOEIL_DEREF_HPP
 
-#ifndef TROMPELOEIL_MOCK_HPP_
-#include "../mock.hpp"
+#ifndef TROMPELOEIL_MATCHER_HPP
+#include "../matcher.hpp"
 #endif
 
 namespace trompeloeil {
